Add *real* snippets.
When you are completing a function/member/pattern/whatever have a
<#...#> in clang's output, snippets can be used. It is automatically launched
(if you have set g:clang_snippets) after completing. To get to the next
<#...#> simply switch to normal mode and push <tab> (in fact it will
select the first <#...#> on the line)

It may be extremely buggy and has only been tested on simple examples.
Note that the plugin wasn't tested with g:clang_snippets disabled, so
the result may have regressed