libclang.py: Introduce option sorting=none
0.34 -> 0.28 sec completion time on (examples/boost.cpp)

This was compared to soring=alpha

No sorting makes sense as vim will automatically sort the results as
soon as we type the first character. So spending a lot of time on
sorting may not be worth it.

In later patches I will try to speed up our own sorting, such that it
becomes less an issue. clang already provides a c function for this.
Though at the moment this function does not work for me (it crashes).