Makes cindex dependency trunk-compatible @162145
The only incompatibility of libclang.py was the new constant
TranslationUnit.PARSE_PRECOMPILED_PREAMBLE.

This fix take freedom of applying corrections and optimizations
to cindex.py, but mantaining a compatible interface.

It corrects the CodeCompletionResults destructor which at @162145
is calling an undeclared function.

It carries optimizations for the CompletionChunk class that do not
exist at @162145.

Both for correction and optimization, patch requests were made to the
Commit List For LLVM's C Front End:
http://goo.gl/ohzrp
http://goo.gl/AfHEa