Basic memory management
Fix the TU "leak" by using a TU pool.
Recent TU's are the last to be deleted. This fix creates a dependency
on the psutil python module (very portable and supported).

User can now set g:clang_memory_percent to give a hint for the maximum
ammount of RAM to be used by TU's. (default = 50)