Fix slowness due to formatting and snippets
Instead of formatting and decorating with snippets all the results
of a completion operation, now it fetches a designated amount of
results for formatting, decorating and add to the completion menu.

The fetching amount is based on pumheight, if it's 0 then a default
value of 50 is used.

This fix the delay due to formatting/decorating a large amount of
results as in 'std::' or 'boost::' for example.