Merge: [cindex.py] Make the use of a compatibilty check explicit
At the moment, we implictly check compatibility between the python
bindings and libclang, as the python bindings will fail to load in
case a method we use in libclang is not available.

This patch makes the use of this compatibility check explicit and
introduces a flag to optionally disable the check. This will allow us to
further harden the compatibility check, but it also gives the user the
possibility to disable the compatibility check to evaluate compatibility
with older libclang versions.

I added documentation that makes clear the python bindings are only
tested with the libclang version they have been shipped with

https://llvm.org/svn/llvm-project/cfe/trunk@163238