Merge: [cindex.py] Replace CachedProperty with our own implementation
This is a very performance critical point for auto completion. The
manual implementation gives a large speedup. As it does not complicate
the code a lot, I figured it is worth the change. If anybody understands
why the CachedProperty is here so much slower, I am very interested in
working on an improvement of CachedProperty.

Formatting time changes from 0.72 to 0.57 seconds.

https://llvm.org/svn/llvm-project/cfe/trunk@172900