libclang: Rework handling of builtin includes
If libclang is installed in some non-default location, it sometimes
fails to find the path to the builtin includes. This will cause an error
when including these headers which again blocks precompiled header
caching. Even though autocompletion often still works, it is annoyingly
slow.

We now check explicitly if the builtin includes are available. If they
are not we try to find them ourselves. Only if this does not work, we
warn the user and ask him to report a problem.