Automatically use libclang if available
If the user did not request anything specific, we first try to use
libclang and only fall back to the clang executable if libclang could
not be found. Error messages are only printed in debugging mode or
if the user explicitly requested libclang.

This patch allows us to use libclang.so as soon as it can be found in
the system library path. No further configuration is required in
.vimrc.