libclang: Allow python snippet functions
Add a python snippet interface. If a snippet engine supports this
interface, snippets are formatted directly from within python. This
makes snippet generation almost zero overhead. If such an interface is
not provided we fall back to the existing interface.

This commit adds a python formatter for the 'dummy' and the
'clang_complete' snippet engine.

This reduces code completion time for examples/boost.cpp from 119 to 90
miliseconds.