Prevent vim from hanging on exit
Until now, vim would hang on exit to wait for the warmup thread or other
completion threads to finish. This can cause an annoying 1-5 seconds
delay when exiting vim. We address this issue by manually closing the
current buffer and then directly killing vim. This is a little hackish,
but we unfortunately do not have any control over running clang threads.
As we close all buffers in advance, the risk that we loose something by
killing vim is minimal. The only remaining annoyance is that vim prints
'Killed' on exit. I doubt there is a lot we can do about this.