Fixing very old bug with g:clang_close_preview set to 1 and pressing key (like semicolon) to close selection ...
When g:clang_close_preview is set to 1 then preview window should close whenever completion is selected.
This worked when enter key was pressed or <c-y>, but when user just wanted to continue the code for example
they pressed a semicolon, then preview window did not close, just the selection menu.

After this it will close even in this scenario (if g:clang_close_preview is set to 1).
Intentionallly in this case the first replacable is not selected (TriggerSnippet() is not called),
this is as it was before in this scenario, so user can continue the writing of the code as they started.
(They already started typing the code after completion and exited from completion selection menu by doing so.)