Show constant value when showing type.
When printing type (g:clang_print_type_key),
also print constant value if known.

Currently macros can only be calculated if you assign them.

In this example, value is not shown when cursor is over MACRO1,
but it is when cursor is over MYCONST1 (14 is shown)
or MYCONST2 (30 is shown):

  #define MACRO1 14
  #define MACRO2(x,y) ((x)*(y))
  const int MYCONST1 = MACRO1;
  const int MYCONST2 = MACRO2(6,5);